requirements.txt
flask>=2.1.2
firebase-admin>=6.0.0
python-dotenv>=1.0.0
protobuf>=4.21.5